All subscription tests passing
